change ubiquity tz to not include the actual tz
ubiquity sometimes goes into UK mode.

which makes no sense.
the server is in germany
the system is in en_US
why the flip would it pick UK as tz, this is utterly mental

exclude the tz information. I really can't be bothered debugging any more
shit with the shit ISOs we are having.