stamp geoip override and use a static port
for oem-config we need to debconf-set in the prepare stage (after install;
before config) as the config stage (first user boot) already runs
oem-config immediately so debconf would get locked.
to bypass this we use a static port and configure it in the prepare stage,
in the config stage we still run the server but without debconfing again.