do not wait still screen. we assert live desktop anyway
wait_still_screen can time out before we are on the desktop, but in the
end we do not really care about it. we care about the desktop appearing.

when we then switch to tty6 we have additional assertions in place to
make sure tty6 is ready to use, so the idle wait is utterly useless really