adjust breeze-cloe again
my previous plan was foiled by the first app that started... discover...
has no widget background at the very top. so. instead. in line of the
same motivation of matching somewhat more aggressively, match the enitre
action portion of the window deocration.

this is slightly less scalable as we now need to have needles for all
permutations of this, but it's still a fair bit better than only
matching the X