add an approximation fallback
if >1 wants the data
but =0 need the data
then fall back to the closest approximation and hope the junit printer code
for the approximation can properly handle the approximate shape.

this is necessary to handle screenshots with no needle. this is different
from a regular screenshot that wasn't meant to match, it is caused
by assert_screen(tags) without any needles matching tags.

this blob e.g.

{:result=>"unk", :tags=>["ubuntu-upgrade-fetcher"], :screenshot=>"upgrade_bionic-29.png", :frametime=>["10.71", "10.75"]}

which is basically a NeedleDetail but without the actual needles field,
since it had no needles.
we can represent this easily enough in a NeedleDetail if we simply init
needles=[].

This could have been represented using a dedicated class BUT, it's easy
enough to do this way.
Should this became a problem in the future we can always unpack this
by simply adding a class that fully matches the blob (match > approx).
However to keep the code dry we'd have to make the class inheritance of
the details more complex, so I am shying away from this for now.

(a NeedleMatchDetail is a NoNeedleMatchDetail AND a NeedleDetail
 a NeedleDetail is a NoNeedleDetail - that's diamond inheritance which
 Ruby doesn't like, so representing it this way is a world of trouble)