adjust sddm needling
sddm previously used incorrect font metrics (and possibly font?), where
it would use dejavu rather than noto. the cause of this is that plasma
never configures fontconfig, so sddm by default would use dejavu sizing.

a new change in neon-settings corrects this problem. to aid in the
transition there are now two sddm needles, one is the preferred full
noto version, and the other the previous semi-noto version

latter is tagged as workaround which should make os-autoinst mark it
matching as soft failure