add a new metadata system to get high level access to global metadata
this is primarily useful to get metadata out of the ISO so we know
what was installed initially.

this data is uploaded by the install core as soon as possible and later
(on success) moved to a metdata dir for archival.
so, for installs the metadata is in ulogs on error and metadata on success

for !install tests they pick up the metadata from the image storage to
seed their initial metadata dir (they may still added more info if
applicable).
they always have metadata (from the install), so even on errors we know
what the image contained

archival of this happens relative, so the metadata is in the main dir