don't match window title
the gradient is messing with openqa something fierce. this is the third
time this particular needle failed because its window size changes all
the time and makes the window title area not match

I don't quite understand why. the title bar seems to stay largely the same
all the time *shrug*