bionic snapd has fewer services, don't treat certain disables as fatal
basically anything but snapd might exist or not, we'll ignore if the
disabling came back false. chances are it did so because the service
doesn't exist.
should this proof to wild an assumption in the future we'll probably want
to do fix it by dealing with the TODO and dynamically list pertinent
services and timers