add bionic support
bionic has a number of differences
- noto font changed a whole lot of characters between the version we have
  in xenial and the one in bionic. this causes spacing changes (usually
  marginal) as well as entire face changes for some characters
- something is bugging out on the live iso for bionic which causes it to
  randomly use incorrect colors for TTY (inherited from grub, which is
  incorrectly themed through live-build; so it may be a side effect of
  that. pending investigation. also changing this means redoing the
  tty needles as all current live needles would have the incorrect colors

to accomodate this we are now adding a bunch of needles which address
some of the graver spacing and character problems to match correctly.

additionally ALL needles are getting their match requirement reduced to 70%
IFF the environment OPENQA_SERIES is bionic. this is meant to make
everything match as much as possible ignoring potential minor delta.
the plan is to slowly go back to 100% an deal with issues as they
appear instead of having to redo all needles all at once.
latter in particular would not work very well because bionic tests do
not have full coverage over all needles until we get closer to the
switch from xenial to bionic