reduce verbosity and reset console after init
outcome is that sddm is correctly running on vt1 but for some reason
switching to it doesn't work.
so, get rid of all excess actions in the upgrade test and instead reset
the consoles after init to make sure they are in a clean state. then
hand over to the plasma_folder test

ideally the folder test should then be able to work. I doubt that's going
to be the case though.
I checked the disk image manually and couldn't reproduce problems with
switching VT. which makes it very odd that somehow the sut remains on
tty6 when it should be on tty1 when switching to x11. additionally the
switch to tty1 seems to be going well on the backend as a casual look
at the upstream code suggests we'd not get activated if the backend failed,
but we definitely get activate_console called.
not sure what to make of any of this :S