try a less invasive (newly added) approach to set the tty in bionic
this is a fairly new feature in os-autoinst, but seems more in line
what we want here. to change the tty but leave the console as is.

however, I really question whether this will help. debug output suggests
that the x11 console should get an activate call. but it decidedly doesn't
get the key input sent.