split lockscreen 5.12 needles and update master needles to new artwork
this tags all 5.12 for neon TYPES that have it and the master needles
unstable

also new idle needle. the lockscreen idles after a while and also initially
so the needle is a replacement for the standard needle. this is a bit meh
but the least invasive way of doing it. this implicitly tests that input
is not lost as we'll basically deal with the idle mode as if it was the
regular UI. has the profound disadvantage that we technically never assert
that the regular UI actually looks what it is supposed to