add crazy sudo caching system
sudo has a passphrase cache on by default meaning we only ever so often
need to auth. this is a problem because we wait 3 seconds for the prompt
every single time something needs to go through sudo.
to improve this we'll cache when an auth has happened on a **tty** console
and only expect an auth if sufficient time has passed.

might be better overall to simply disable password requirement in
first_start.pm so sudo will simply work without us having to wait for auth
anywhere