limit concurrent installation tests via jenkins resources
we are having continued problems with installation tests semi-randomly
failing. the working theory is that there are IO problems caused by
concurrently running install tests using too many resources. as a quick
fix we'll serialize the installations, which should, in theory anyway,
reduce the number of flaky runs to zero