change locking to happen before claiming the node
before we'd claim the node,, enter a stage and then lock the serializer.
while that is fine from a serialization pov it's bogus from a node
pov as we then sit there in the build slot of the node without doing
anything.
so, instead claim the lock and then the node. this way we are basically
"queued" up until we have the lock at which point we then also claim our
build slot