attempt to detect the idle screen more reliably
this is a bit problematic as the idle screen contains entirely mutable
data which is always different, so we have a hard time matching the text.
at the same time only matching the background is silly too as this means
we may match on the desktop still.

make the needle area cover the mutable data but with super low match level.
my hope is this will be sufficiently strict to only match the idle screen
but not the desktop.

if all else fails we may opt to either give the lockscreen a different
color, or (more likely) introduce a sleep after the locking