adjust the idle needle to cover just about the entire screen
this needle is super tricky. we need to differentiate from the desktop
(which is largely blue) AND we need to differentiate from !idle lockscreens
(which is largely blue). to stand a better chance we now assert the
full width and the full height of the screen via areas as well as just how
much overall background we'd like to see. this should differ enough from
other lockscreen states to be distinct, if it still isn't distinct enough
from the desktop we may need to use a different color though