bump the base image max size to 30G
the massive upgrade test is too large to fit into 20G and we can't easily
upgrade to a larger overlay because we'd need to resize the file system
and calamares wisely puts the swap at the end of the disk, so we have to
move the swap, then umount, then resize, then remount/reboot. fairly
extensive. while that is nice and possible should happen in the future
for now I'd like a straight forward solution.