fix archiving when install tests fail early
when an install test fails before it reaches the original metadata
retrieval (e.g. broken uefi) the metadata dir may be empty, therefore
we cannot expect it not to be empty.
instead we'll allow empty, do the rest of the post-build archiving and
then explicitly attempt to ls * which should then end in error if sh
cannot expand * to anything.

without this the metadata archival aborts in error and none of the steps
after it are actually reached. also, moving the archival to the end made
similiarly no sense since the steps such as chown may also fail, so we'd
then skip archiving on account of unrelated error. archiving and split
assertion seems the best way really