properly set iso relation + disable concurrent builds
calling properties in a jenkinsfile overrides whatever was set in the
xml. this is a problem as running a job that does it will undo the
concurrent disablement done in the tooling templates. so, we also disable
here. tooling is the default value. the values from the jenkinsfile only
get applied after first run

also properly init openqa_series to xenial unless otherwise defined
(which it should be moving forward FTR)