do not override properties in jenkinsfile
we set properties in our pangea-tooling templates, and those are
persistently overwritten by a properties() call. it's unclear why, looks
like a bug though. in any case we don't really need properties here anymore
with the pangea-tooling having a good way to manage properties of the
jobs already.

leave behind a warning so we don't accidentally re-introduce properties()

if we ever need them again, great care must be taken that running a job
with properties() doesn't unexpectedly lose the XML defined properties