switch encrypted home upgrades on by default
this test makes sure encrypted homes are still encrypted and accessible
after upgrades. this was opt-in while I was still refining it. it's working
perfectly now and can be used for all upgrade tests.

upgrade tests take a fair amount of time and resources, so keeping them
very compact is a good idea.