add a completely new partitioning-only test for calamares
recently kpmcore got a large change which broke a whole bunch of stuff.
this test switches cala into a partitioning only mode and then uses the
features erase, replace, resize, and wipe in that order. this happens all
on a GPT table. for the grand finale we then clear the table and change to
MBR and do a simple manual partitioning.

none of this asserts the outcome is actually correct (which is a bit tricky
to do anyway). instead we primarily care about the partitioning failing
entirely.