implement a somewhat cheap trick to get some useful info from drkonqi
this only works when the window is already up which it may or may not be
depending on a number of factors.

this changes simply switches to the dev info tab. it'll likely not
show anything useful though because the window is super small and the
backtrace textfield super small

better than nothing tho