drop screenlock disabling
according to the upgrader code it should disable the locker already.
if this isn't the case it should be fixed.

manually disabling the locker here in the test only hides the problem, the
user may still get a lock and then be unable to unlock because of runtime
asset loading problems

T9534