set static colors for wallpaper and lockscreen in first_start after install
also let the existing code bailout if it detects that the colors were
set. this is purely transitional until the images have all been redone
with the new first_start. until then we'll record a soft failure.

this completely solves the changing wallpaper problem, previously only
the plasma test had it solved because of the color change code. now that
it is in first_start all tests will have static backgrounds except for
live testing