enable sysrq + add special sysrq debugging when stuck on plymouth
this is special error handling.

we have persistent problems that reboots get stuck. previous debugging
attempts via journald didn't give useful results because according to the
journal the system is shutting down as expected. It's clearly stuck on
something though.

so... to debug this we'll try to get as much out of the kernel as possible
and then crash the system for good measure.

additionally we'll now enable sysrq all the time since ubuntu's default
security would disable most of the debugging commands. since we have no
tests relying on the security being present we may as well just disable it