fix relogging by using the actual console activation logic
screw manually relogging on bionic. it's highly defective

instead add a needle for when encrypty is logged in and actually drive th
relog through select_console.

we now set the username/password for the testapi accordignly

- go tty6
- verify
- logout
- reset
- go to x11
- change testapi to encrypty
- go to tty6 [test api logs in]
- verify
- logout
- reset
- go to x11
- reset testapi to user

this way we avoid having to manually drive the login, get rid of the sleep
for being logged in, and this should be state-safe for bionic as
we actually fully reset console states now