reduce wait on switching to x11
wait_still_screen is internally nothing but a sleep (a longish one), we
don't really need a long wait for the VT switch. 2 seconds should be
plenty. if not the run probably should have its timing scale increased
anyway (i.e. make everything have more leeway because the server is
slow or whatever)