don't wait screen after livetest setup
the original assertion is already that we are on the desktop. then we
switch to a tty, login, do stuff, go back to x11 AND wait 2 seconds.
if the session isn't ready by the time we come back from all that it's
unlikely to pass responsiveness expectations the way it is and needs
time scaling applied to account for the slowness