reduce the wait time for sudo a bit and stop waiting
this takes screenshots as quickly as it can to decrease time between
state change and matching making everything faster if a prompt appears.
at the same time we'll reduce the overall time the prompt has to appear
from 3 to 2 seconds to make the no-prompt scenario faster as well.

ultimately this is a bit sucky because of the fact that we cannot really
tell the two states apart, so we end up waiting when no prompt is going
to appear anyway, or we wait not long enough when a prompt needs to appear.

it may be worth considering to switch the qa images to not cache sudo at all
that we we always know a prompt must appear. then again always having
to wait for the prompt and entering password may end up slower with
many sudo calls. all very meh