unify live session reboots and force reset the system if necessary
ISO still get stuck on reboot sometimes. I've tried for days to get this
properly fixed up now and failed miserably. I am almost entirely certain
it has to do with dodgy unit order of casper.service, but I can't figure
out what exactly is wrong or what it should be like. ultimatley
casper.service is a bit garbage because it needs to run some time between
plymouth starting and final happening, when exactly is hard to find out
and also this is shitty to test since the only way to semi reliably
reproduce it is via openqa.... so, instead simply force a reset of the
power if the sytem didn't properly reboot after ~20 seconds of us hitting
enter on the 'hit enter' screen.
it sucks massively from a test reliability POV, but it's the only way I see
to prevent flakeyness