revise the ubiquity test to run twice
https://bugs.kde.org/show_bug.cgi?id=398752 seems to be caused by having
a pre-existing partition table. this is a bit tricky to test with ubiquity
so we'll do a fairly dirty approach by running the installation twice and
force resetting the system in between. this should give the second run
the same situation that causes the bug. this should probably stay forever
more to make sure this doesn't regress.

unfortunately this makes the test take twice as long