explicitly boot from cdrom when installing
we expect the cd to get 'ejected' before rebooting, if that is not the case
we'll consider this a problem anyway. I hope this works as expected
anyway.

this change is meant to facilitate a power reset when installing via
ubiquity twice. on the boot after reset we come up with the disk already
even though we didn't eject the cd, so I am expecting qemu simply boots
from the hard disk because it has data and is ordered before the cdrom, so
if we force cdrom to be preferred we'd ideally always boot from cdrom in
installation tests except for when we eject it (which we do for the final
reboot)