rejigger the ubiquity partitioning test
while this still sucks because it is part of ubiquity test proper, this is
no longer part of the standard installation test.

instead the partitioning test only runs when the correct var is set but it
is still part of the regular installer code.
additionally the BOOTFROM setting is now also conditional to the var, as
unfotunately this behaved as I feared and prevents disk booting in
first_start, so we cannot use this var for regular installation tests where
we need to boot from disk