prevent generic details from factorizing
with latest master I am seeing tests (supposedly) failing as you would
expect, but then the output json contains generic details with no
worthwhile information. prevent such details from factorizing.

regardless of other defects that may play into this result there is not
a single reason where factorizing a worthless detail is something we need
or want to do. instead ignore them entirely. this is a bit shoddy mapping
but TBH the detail entry in of itself is too