carry isotovideo exit code forward
new os-autoinst indeed returns !0 when a test fails. save the return code,
then convert to junit result (if available anyway). in the event that junit
conversion returns on error (which it generally doesn't and shouldn't)
explicitly exit with the same code as isotovideo

i.e. if isotovideo came back !0 run.rb will do the same. always.