repair post-install tests
kind of anyway. with the new os-autoinst snapshotting is a bit of a mess
and so is the raid/ directory. it's all in all much harder to get
efficient post-install tests going because keep_hdd mode (which is what
we use for post-install) no longer creates a new "profile" without also
creating a new hard disk, but instead loads the qemu_state of a previous
profile. IOW a post-install test would load the qemu drives of the
install test it is derived from. that technically is lovely but for two
things. a) snapshotting seem to result in tianocore not booting after
the snapshot. no clue why but that breaks uefi. b) we don't actually
want the neon.iso be required, but it would be because the cd drive is
based on it and qemu wouldn't start if it can't find the backing iso file.

as a somewhat reliable fix for this the raid recycling is changed a bit.

- install tests now disable snaphots, this prevents uefi from getting
  messed up for unknown reasons to do with snapshotting. plus we don't
  use snapshots anyway, so there's no reason why it would be enabled.
- the !installation test setup now also copies the pflash drives (we
  only test on uefi, so the pflash drives are always present and necessary)
- the qemu_state.json is is now moved into the raid/ dir after the test
  and subsequently picked up by !installation tests and copied into their
  wok
- the qemu_state.json is further mangled to entirely drop the cd drive and
  by extension the requirement on neon.iso

this should make !installation tests work once again. it's a bit of a mess
though, conceptually speaking this is really far removed from how an
ordinary openqa install would do this :/