increase the sleep time in an effort to see if it is a load problem
I cannot reliably reproduce this locally anymore, but on the test server
we still have the test flake on the okular context action. logs suggest
that plasma is still loading delayed plugins at that point in time. at the
same time the positinochanged signal still do not arrive. this bug is super
stupid -.-