reduce sleep overhead
may or may not still pass this way. if the problem isn't a general
load issue it may be that event just happen to quickly?
if the qt output is to be trusted there definitely isn't a second event
being handled. so either qt ignores the two additional moves, or,
qemu/vnc is not doing them