fix manual partitioning default table type
this was incorrectly mbr before, but since the system is UEFI it should
always have defaulted to gpt. I totally missed that originally.

update the expectation to be gpt by default and rename the needle
accordingly. calamares now also behaves as one would expect and comes up
with gpt by default when creating a new table on an uefi system