add a sleep while partitioning
this is a try to prevent the partitioning failing on obscure fdisk
errors without useful error output. I believe this is because of a race
when initializing the modules, but find it hard to proof that theory.
if a sleep really fixes the problem in many runs then we can be fairly
certain that it is a race.